% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAssayVariance.R
\name{addAssayVariance}
\alias{addAssayVariance}
\title{Add assay Noise to CDF plot}
\usage{
addAssayVariance(
  apt,
  apt_median,
  assay_var = NULL,
  mad = NULL,
  matrix_type = c(NA, "plasma", "serum"),
  ...
)
}
\arguments{
\item{apt}{\code{character(1)}. Which feature to add. This queries the
\code{splyr::mLod} object to obtain the median absolute deviation
(MAD) for that aptamer. Ignored if \code{mad} is passed directly.}

\item{apt_median}{Numeric. The median value for the empirical values
to be plotted against. This is used to create a random normal of the
same median but with the assay variance from the median absolute
deviation (pinned shift).}

\item{matrix_type}{Serum or Plasma.}

\item{...}{Additonal arguments passed to \code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDF()}}.}
}
\description{
Add a CDF of the estimated assay variance via the Covance QC samples.
}
\author{
Stu Field
}
