% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCVbyGroup.R
\name{calcCVbyGroup}
\alias{calcCVbyGroup}
\title{Calculate CVs by Group}
\usage{
calcCVbyGroup(data, group.var)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object.}

\item{field}{Character. The group variable used to split the data.}
}
\value{
A tibble with rows as the \%CV for each group, and
columns as analytes.
}
\description{
Calculate CVs by a grouping variable, e.g. \code{SiteId}.
}
\examples{
calcCVbyGroup(sim_adat, "SiteId")
}
\author{
Stu Field
}
