% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPairwiseTables.R
\name{calcPairwiseTables}
\alias{calcPairwiseTables}
\title{Pairwise Comparison Tables}
\usage{
calcPairwiseTables(dat, group.field, comps = NULL, ...)
}
\arguments{
\item{dat}{A data set or "soma.adat" object containing RFU values
and a \code{group.field} column containing a vector of levels
to compare in each data split. The levels of this field should
be \code{>2}.}

\item{group.field}{Character. The name of the column in the
ADAT containing the grouping information to split the data into
pairwise groups.}

\item{comps}{A n x 2 matrix of the group name comparisons you wish to make.
Defaults to all pairwise \code{choose(n, 2)} comparisons.}

\item{...}{Arguments passed to \code{\link{createTestsList}}}
}
\value{
A list of statistacal tables, each the result of a
call to the specified function on a pairwise split of the data.
}
\description{
This is a pairwise wrapper for \code{\link{createTestsList}}, which
calculates a statistics table for each pairwise group set that
can be calculated from a given set of factor levels. There are
typically \code{choose(n, 2)} pairwise comparisons to make and
thus \code{Nchoose2} tables are generated.
}
\examples{
tmp <- sample.adat
tmp$newGroup <- rep(head(LETTERS, 4), each = 5)
tests <- calcPairwiseTables(tmp, group.field = "newGroup", test = "t", do.log = TRUE)
names(tests)
}
\seealso{
\code{\link{createTestsList}}, \code{\link[utils]{combn}}, \code{\link[SomaClassify]{createTrainingData}}
}
\author{
Stu Field
}
