% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPercentileTable.R
\name{calcPercentileTable}
\alias{calcPercentileTable}
\title{Calculate percentiles}
\usage{
calcPercentileTable(x, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{x}{A \emph{named} list of vectors.}

\item{probs}{Numeric. A vector of probabilities in \link{0,1}.}
}
\value{
A table of (usually CVs) split by percentile.
}
\description{
Calculate percentiles of a series of vectors and put
into tablular form.  By default, the 5th, 25th, 50th, 75th,
and 95th quantiles are generated, though any quantiles can be displayed.
Often used for new matrix characterization of CVs.
}
\examples{
x <- lapply(seq(50, 250, by = 50), function(x) rnorm(100, mean = x))
names(x) <- head(LETTERS, length(x))
calcPercentileTable(x)
}
\seealso{
\code{\link{quantile}}
}
\author{
Stu Field
}
