% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRsquared.R
\name{calcRsquared}
\alias{calcRsquared}
\title{Calculate R-squared}
\usage{
calcRsquared(model)
}
\arguments{
\item{model}{A linear model of class "lm".}
}
\value{
The R-squared of a model.
}
\description{
Returns the residual sum of squares of a model
}
\details{
Details
}
\author{
Stu Field
}
