\name{clone-methods}
\docType{methods}
\alias{clone-methods}
\alias{clone,RLBigData-method}
\alias{clone,RLResult-method}
\title{Methods for function\code{clone}.}
\description{
Methods for function \code{clone}. See \link{clone}
for details.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "RLBigData")}}{

}

\item{\code{signature(object = "RLResult")}}{

}
}}
\keyword{methods}
\keyword{classif}
