% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCVBands.R
\name{computeCVBands}
\alias{computeCVBands}
\alias{calcCVbands}
\title{Compute CV Bands}
\usage{
computeCVBands(rfu.range, cv, iter = 1000, conf = 0.95)

calcCVbands(rfu.range, cv, conf = 0.95)
}
\arguments{
\item{rfu.range}{RFU (linear) values for which CV bands will be calculated
Best if this is something like seq(100,1000,10).}

\item{cv}{The CV as decimal (i.e. 5\\% CV=0.05)}

\item{iter}{Integer. Number of iterations of random adjustment.}

\item{conf}{Numeric. The empirical confidence interval as decimal (i.e. 95\\% CI=0.95).}
}
\value{
A data frame with three columns:\cr
\item{rfu}{the rfu values from input parameters}
\item{upper}{the corresponding upper eCI band for each RFU}
\item{lower}{the corredspoding lower eCI band for each RFU}
}
\description{
Calculate random variations for the passed RFU range using
the specified CV level. Computation is repeated \link{iter=}
times, and the empirical CI limits returned.
}
\section{Functions}{
\itemize{
\item \code{calcCVbands()}: Calculate confidence interval bands. Analytical estimation.

}}
\note{
Empirical estimation.
}
\examples{
\dontrun{}
}
\author{
Kirk DeLisle

Stu Field
}
