% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConcordTable.R
\name{createConcordTable}
\alias{createConcordTable}
\title{Create Concordance Table}
\usage{
createConcordTable(dat)
}
\arguments{
\item{dat}{The data frame containing data.}
}
\value{
A data frame.
}
\description{
Description
}
\examples{
createConcordTable(dat)
}
\author{
Stu Field
}
