% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNormalizationSummaryTable.R
\name{createNormalizationSummaryTable}
\alias{createNormalizationSummaryTable}
\title{Create Normalization Summary Table}
\usage{
createNormalizationSummaryTable(dat)
}
\arguments{
\item{A}{median normalizaed and calibrated ADAT, with multiple
plates that have been calibrated into one data frame ("soma.adat").}
}
\value{
A summary table of the calibration results
}
\description{
This function summarizes the normalization and calibration metrics
into a table. It includes the plate normalization if performed
and the hyb and med normalization scale factor ranges. This
function is primarily used in the 3 plate reproducibility study
experimental desin, but could be used for any normalized ADAT.
}
\examples{
\dontrun{
createNormalizationSummaryTable(dat)
}
}
\author{
Stu Field
}
