% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestsList.R
\name{createTestsList}
\alias{createTestsList}
\title{Create Various Stat Test Tables}
\usage{
createTestsList(
  dat.list,
  test = c("t", "wilcox", "ks", "cor", "lr", "kw", "mackwolfe"),
  which.paired = numeric(0),
  ...
)
}
\arguments{
\item{dat.list}{A named list of data frames containing SOMAmer data.}

\item{test}{Matched String. Which statistical test to perform.}

\item{which.paired}{which \code{dat.list} entries are paired analyses
(user must order properly)}

\item{...}{Additional arguments passed through to the eventual "test fun",
e.g. \verb{\link[stats]\{t.test\}}.}
}
\value{
A named list of class "data.frame" tables, one for each entry of
\code{dat.list}.
}
\description{
Creates appropriate statistical output tables
for a given named list of data frames containing SOMAmer data.
Available tests are:
\describe{
\item{\code{t}}{t.test}
\item{\code{wilcox}}{mann-whitney}
\item{\code{ks}}{ks.test}
\item{\code{lr}}{log ratio}
\item{\code{cor}}{pearson or spearman}
\item{\code{kw}}{kw.test}
}
}
\examples{
age_dfs <- split(sample.adat, sample.adat$TimePoint) # split by Young/Old
sapply(age_dfs, function(x) table(x$SampleGroup))
ks <- createTestsList(age_dfs, test = "ks", response = "SampleGroup")
t <- createTestsList(age_dfs, test = "t", response = "SampleGroup")
}
\seealso{
\code{\link{get}}
}
\author{
Stu Field
}
