% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUrineNormalizationSummaryTable.R
\name{createUrineNormalizationSummaryTable}
\alias{createUrineNormalizationSummaryTable}
\title{Create Normalization Summary Table for Urine}
\usage{
createUrineNormalizationSummaryTable(dat)
}
\arguments{
\item{dat}{A 3-group normalized ADAT, with multiple
plates that have been normalized into one data frame ("soma.adat").}
}
\value{
A summary table of the normalization results
}
\description{
This function summarizes the unrine normalization metrics
into a table. It includes the plate normalization
and the hyb and med normalization scale factor ranges, for the "S2" group.
}
\examples{
\dontrun{
createUrineNormalizationSummaryTable(dat)
}
}
\author{
Stu Field, Eduardo Tabacman
}
