% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dims.R
\name{data_dims}
\alias{data_dims}
\title{Summarize dimensions of list of data frames}
\usage{
data_dims(x)
}
\arguments{
\item{x}{A list of objects to be evaluated using \code{dim}.}
}
\value{
A list of dims each of the dimensions in the
elements of the data list.
}
\description{
Returns the dimensions of the objects contained within a list of
objects, typically a list of data frame objects.
}
\examples{
tmp <- lapply(c(2, 3, 6, 8, 12), function(x) matrix(1:24, ncol = x))
names(tmp) <- head(letters, 5)
data_dims(tmp)
}
\author{
Stu Field
}
