% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denormalizationWrapper.R
\name{denormalizationWrapper}
\alias{denormalizationWrapper}
\title{Complete denormalization wrapper}
\usage{
denormalizationWrapper(adat)
}
\arguments{
\item{adat}{A fully normalized adat to be processed (multiple runs, usually),
should have a \code{PlateId} column.}
}
\value{
A list of ADATs consisting of:
\item{raw }{Raw ADAT }
\item{hyb }{Internal hybridization }
\item{imed}{Internal median normalized (of only Calibrators and Buffers;
as per V4 scheme).}
\item{plt }{Plate scaled}
\item{cal }{Calibrated}
\item{med }{Median normalized}
}
\description{
Complete wrapper around functions and processes necessary
for full denormalization of ADATs
\strong{Note:} we assume standard arguments for the normalization process
have been used. In particular, that the steps are:
hybNorm -> internal med norm of Calibrator|Buffer ->
Plate Scaling -> Calibration -> median normalization of the rest of the samples
}
\examples{
\dontrun{}
}
\references{
SomaLogic Platform BI
}
\seealso{
\code{\link{normalizationWrapper}}
}
\author{
Stu Field, Eduardo Tabacman
}
