% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-package.R
\docType{package}
\name{devel-package}
\alias{devel}
\alias{devel-package}
\title{devel: R Development Scripts}
\description{
This "package" is the development package. Do not attempt to load it as you would a typical package. It contains a loose testing ground of functionality that, if deemed sufficiently useful, will eventually be migrated into an existing package.
}
\seealso{
Useful links:
\itemize{
  \item \url{www.somalogic.com}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
