% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicatedIndex.R
\name{duplicatedIndex}
\alias{duplicatedIndex}
\title{Determine Duplicated Entries}
\usage{
duplicatedIndex(x)
}
\arguments{
\item{x}{A vector containing (possibly) duplicated entries.}
}
\value{
A vector of the indices of the duplicated entries.
}
\description{
Calculates the index/indices of the duplicated entries
of a vector (string or numeric).
}
\examples{
duplicatedIndex(c(1, 1:10))
duplicatedIndex(c("a", "a", "b","c"))
}
\seealso{
\code{\link[base]{duplicated}}
}
\author{
Stu Field
}
