\name{emWeights-methods}
\docType{methods}
\alias{emWeights-methods}
\alias{emWeights,RecLinkData-method}
\alias{emWeights,RLBigData-method}
\title{Methods for Function emWeights}
\description{
  Calculates EM weights for record pairs. See \code{\link{emWeights}} for
  details.
}
\section{Methods}{
\describe{

\item{\code{signature(rpairs = "RecLinkData")}}{
}

\item{\code{signature(rpairs = "RLBigData")}}{
}
}}

\author{Andreas Borg, Murat Sariyar}
\keyword{methods}
\keyword{classif}
