\name{epiWeights-methods}
\docType{methods}
\alias{epiWeights-methods}
\alias{epiWeights,ANY-method}
\alias{epiWeights,RLBigData-method}
\title{Methods for Function epiWeights}
\description{
  Calculates EM weights for record pairs. See \code{\link{emWeights}} for
  details.
}
\section{Methods}{
\describe{

\item{\code{signature(rpairs = "ANY")}}{
}

\item{\code{signature(rpairs = "RLBigData")}}{
}
}}

\author{Andreas Borg, Murat Sariyar}
\keyword{methods}
\keyword{classif}
