% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCalibrationQC.R
\name{evaluateCalibrationQC}
\alias{evaluateCalibrationQC}
\title{Evaluate calibration via QC}
\usage{
evaluateCalibrationQC(cal.data, QCref = NULL)
}
\arguments{
\item{cal.data}{A \code{soma.adat} object of calibrated data containing QC
samples of interest.}

\item{QCref}{A data.frame object with rows as the SOMAmers and columns as
vectors of references for each of the QC lots. As a default, the
\code{QCref} is calculated according to an internal reference. If other
references are to be used, see \code{\link{calcQCref}}, or create one
yourself.}
}
\value{
A data frame containing the absolute values of the log-transformed
ratios of the QC-reference / qc-value for each plate and QC lot. These
values are then used to evaluate calibration performance. Median CDFs of
each of the runs should be less than 0.1(?) (\strong{refer to SOP}).
}
\description{
This function calculates the minimum ratio of the QC samples to their
respective reference, or expected value.
}
\details{
For each QC-sample on each plate, a ratio of the QC-reference/QC-sample is
calculated, for typically 4 samples per plate (duplicate samples x 2 QC
lots). The absolute value of the log of this ratio is calculated and the
\emph{minimum} of the 4 transformed values is returned. The QC reference can
be either calculated internally (default) or passed via the
\code{\link{calcQCref}} argument. If calculating externally, the
QC-reference must be made from calibrated QC-samples from the same runs used
to create the reference to which the current data was calibrated (i.e. for
internal calibration, said data is itself).
}
\examples{
\dontrun{
evaluateCalibrationQC(calibrated.data)
}
}
\references{
Darryl Perry
}
\seealso{
\code{\link{filterAdat}}, \code{\link{calcQCref}}, \code{\link{getAptamers}}
}
\author{
Stu Field
}
