% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilutionNormalizeNPL.R
\name{fit.npl}
\alias{fit.npl}
\alias{print.npl}
\title{Fit an n-Parameter Logistic Model}
\usage{
\method{fit}{npl}(data, npar = 4, do.log = TRUE, ...)

\method{print}{npl}(x, ...)
}
\arguments{
\item{data}{A data frame consisting of 2 columns of data, that \emph{must}
be named "x" and "y". In the context of fitting logistic models for dilution
normalization, "x" and "y" correspond to the dilutions and RFU values respectively}

\item{npar}{Numeric. The number of parameters in the logistic model fit.
Currently either 4 or 5 are supported, with 4 being the default}

\item{do.log}{Logical. Should the y-values (RFUs) be log10-transformed
proir to fitting?}

\item{...}{Additional arguments passed to \code{\link[stats]{optim}}, OR
arguments passed to \code{\link{curve}} if used in the S3 plot method, OR
arguments passed to the generic S3 \code{print} methods.}
}
\value{
A list containing:
\item{data}{The original data for the fitted points (after outlier removal)}
\item{minimum}{The minimum value of the objective function}
\item{pars}{The 4 (or 5) parameter estimates for: L, U, k, b, and s (if 5PL fit)}
\item{iterations}{The number of iterations needed for convergence}
\item{convergence}{Was convergence reached, if not, the corresponding error code}
\item{fitted.y}{}
\item{newx}{}
\item{newy}{}
\item{residuals}{The residuals of the model fit}
\item{pt.infl}{The inflection point of the model}
\item{limits}{The linear limits of the logistic curve relative to the RFU axis}
\item{stats}{Fitting statistics: Goodness of Fit, Standard Error, p-value, and AIC}
\item{npar}{Number of parameters in the logistic fit}
\item{outliers}{Were outliers detected and removed prior to the fitting?}
}
\description{
Description ...

A generic print call for objects of the "npl" class.
Objects are created by a call to \code{\link{fit.npl}}.
}
\details{
Currently only supports 4 or 5 parameter models
}
\examples{
\dontrun{
npl <- fit.npl(dilution.data, npar = 4, method = "BFGS", control = list(maxit = 5000))
npl
npl \%>\% get_pars
npl \%>\% plot
}
}
\seealso{
\code{\link[stats]{optim}}

\code{\link{cat}}
}
\author{
Stu Field
}
