% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSingleSampleReference.R
\name{generateSingleSampleReference}
\alias{generateSingleSampleReference}
\title{Create Single-Sample Reference}
\usage{
generateSingleSampleReference(
  adat,
  initialRef = NULL,
  effect_cut = 2,
  recenter = FALSE,
  maxrelerr = 0.01,
  maxiter = 20
)
}
\arguments{
\item{adat}{the data to use for reference building.}

\item{initialRef}{a reference to kick-start the process.
If NULL (default), use all of the adat samples}

\item{cutoff}{the limit to use in the call to ssANML (default: 2),
to decide which analytes get dropped}
}
\value{
a reference object (see generateRefParams below)
of the final normalized data
}
\description{
Create a reference from un-median normalized (but calibrated) data
using single-sample adaptive maximum likelihood normalization (ssANML)
We start with an initiaReference and an adat (in particular, we use
Covance-Boise for the initialReference, and all of Covance ), then use ssANML
on all the samples and rebuild references from them. Rinse, repeat until SFs don't change.
}
