% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_availability_matrix}
\alias{get_availability_matrix}
\title{Extract the availability matrix from the biological Parameterfile (*.prm)}
\usage{
get_availability_matrix(model_path, filename)
}
\arguments{
\item{model-path}{}

\item{filename}{of the parameterfile}
}
\description{
This function extacts the availability matrix from the biological Parameterfile
}
\examples{
get_availability_matrix(model_path = "Z://Atlantis//ATLANTIS NSmodel", filename = "NorthSea_biol_fishing_final.prm")
}

