\name{getDbFile-methods}
\docType{methods}
\alias{getDbFile-methods}
\alias{getDbFile,RLBigData-method}
\title{Get path to database}
\description{
  Utility method to get the path of the database file of a
  record linkage data or result object.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "RLBigData")}}{
}
\item{\code{signature(object = "RLResult")}}{
}
}}
\keyword{methods}
\keyword{file}
