% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDupIdxList.R
\name{getDupIdxList}
\alias{getDupIdxList}
\title{Get Duplicated Index List}
\usage{
getDupIdxList(x)
}
\arguments{
\item{x}{A vector of values with potentially duplicated values.}
}
\value{
A list of indices split by duplicated value of any duplicated values.
}
\description{
Calculate the indices of the duplicated values of a vector
and return a list corresponding to the duplicated groups.
}
\examples{
x <- rep(head(LETTERS, 10), c(1,1,2,3,1,2,3,1,2,1))
x
table(x)
getDupIdxList(x)

getDupIdxList(1:10)         # no duplicates (empty list)
getDupIdxList(LETTERS)      # no duplicates (empty list)
getDupIdxList(c(1, 1:10))   # 1 duplicate
}
\seealso{
\code{\link{duplicatedIndex}}
}
\author{
Stu Field and Yolanda Hagar
}
