\name{getPairs-methods}
\docType{methods}
\alias{getPairs-methods}
\alias{getPairs,EMWeights-method}
\alias{getPairs,RecLinkData-method}
\alias{getPairs,RLBigData-method}
\alias{getPairs,RLResult-method}
\title{Retreive record pairs}
\description{
  Retreives record pairs from a record linkage project or result set.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "EMWeights")}}{
describe this method here~~

}

\item{\code{signature(object = "RecLinkData")}}{
describe this method here~~
}

\item{\code{signature(object = "RLBigData")}}{
  \usage{
 \S4method{generic}{signature_list}(argument_list)
 %  \S4method{getPairs}{RLBigData}(filter.match = cc("match", "possible", "non-match"),
%    single.rows = FALSE)
  }
  }


\item{\code{signature(object = "RLResult")}}{
describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
