% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRFUdata.R
\name{getRFUdata}
\alias{getRFUdata}
\title{Get Data Matrix of RFU Data Only}
\usage{
getRFUdata(x)
}
\arguments{
\item{x}{A data frame of class \code{soma_adat}.}
}
\value{
A data matrix of the RFU data only.
}
\description{
Simple function to subset the feature data (RFU columns)
and remove any meta data columns that might be present.
Some functions require a simple data matrix of measurements
to be passed, this this function strips non-proteomic data
and recasts the data frame to a matrix class.
}
\examples{
class(sample.adat)
dim(sample.adat)
class(getRFUdata(x))
dim(getRFUdata(x))
}
\seealso{
\code{\link{data.matrix}}
}
\author{
Stu Field
}
