% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSSmenu.R
\name{getSSmenu}
\alias{getSSmenu}
\title{Get SomaSciences Menu Content}
\usage{
getSSmenu(menu.ver)
}
\arguments{
\item{menu.ver}{Character. Which SomaSciences menu version to select}
}
\value{
A character vector of SeqIds of the SomaSciences menu of interest
}
\description{
Get a list of SeqIds corresponding to a particular
SomaSciences content version. This function depends on
\code{\link[SomaPlyr]{SSmenu}}.
}
\examples{
getSSmenu("v1.1k_v01") \%>\% head(10)
}
\seealso{
\code{\link[SomaPlyr]{SSmenu}}
}
\author{
Stu Field
}
