% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAxes.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color Hue for ggplot}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{the number of colors to use}
}
\value{
a vector of colors to pass into ggplot
}
\description{
Function to get the (default) ggplot colors, in case one needs
to match some default plot
}
