% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsObjects.R
\name{lsObjects}
\alias{lsObjects}
\title{List Object Size in an Environment}
\usage{
lsObjects(env = .GlobalEnv, units = "MB")
}
\arguments{
\item{env}{The environment to be searched.}

\item{units}{The units of the object size in the output data frame.}
}
\value{
A tibble ordered first by the object size.
}
\description{
Creates a data frame of all the objects in a specified environment,
and returns their object size (in KB), the class
of that object, and the dimension of the object (unless object is a function).
}
\examples{
lsObjects()
}
\seealso{
\code{\link[base]{ls}}, \code{\link{switch}}
}
\author{
Stu Field
}
