% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss.R
\name{lss}
\alias{lss}
\title{List Env Functions & Objects (deprecate?)}
\usage{
lss()
}
\arguments{
\item{env}{Which environment is to be searched, defaults to \code{.GlobalEnv}.}
}
\value{
A list of objects separated by object classes, and print a summary
of the numbers of each object by class.
}
\description{
This function separately lists the functions and objects in global memory
(default).
}
\examples{
lss()
}
\seealso{
\code{\link{ls}}
}
\author{
Stu Field
}
