% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMetaData.R
\name{mergeMetaData}
\alias{mergeMetaData}
\title{Merge Meta Data}
\usage{
mergeMetaData(adat, meta, ..., fix.atts = FALSE)
}
\arguments{
\item{adat}{The existing adat to which the meta data is to be merged.}

\item{meta}{Can be either:
\enumerate{
\item A data frame of additional meta data, with at
least one common column name to match the samples OR
\item A file name (and path) pointing to where to get a
"*.csv" file containing meta data (this option uses a call
to \code{\link[utils]{read.csv}}).
}}

\item{...}{Additional arguments passed to \code{\link[dplyr]{left_join}}.}

\item{fix.atts}{Logical. Should the attributes of the merged adat be fixed
by \code{\link{createChildAttributes}} or left broken?}
}
\value{
An adat data frame with extra columns corresponding to the meta data
added and indexed to the appropriate sample IDs.
}
\description{
Merge additional meta data to an existing adat based on unique row sample
identifiers. This is a wrapper around the existing R function
\code{\link[dplyr]{left_join}}.
}
\examples{
set.seed(101)
new <- tibble::tibble(SampleUniqueID = sample(sample.adat$SampleUniqueID),
                      NewData = rnorm(20))
mergeMetaData(sample.adat, new)                         # use default 'by'
mergeMetaData(sample.adat, new, by = "SampleUniqueID")  # same
}
\seealso{
\code{\link[dplyr]{left_join}}, \code{\link[utils]{read.csv}}, \code{\link{createChildAttributes}}
}
\author{
Stu Field
}
