% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyAptContent.R
\name{modifyAptContent}
\alias{modifyAptContent}
\title{Modify Aptamer Content}
\usage{
modifyAptContent(adat, menu.ver = "v1.1k_v01")
}
\arguments{
\item{adat}{The adat to be modified.}

\item{menu.ver}{Numeric or Character. Which SomaSciences menu version to
use. The default is \code{menu.ver="v1.1k_v01"} which corresponds to the
1129 menu. Should match one of the names of the \code{SSmenu} object.}
}
\value{
An adat with modified aptamer content.
}
\description{
Modify aptamer content of a data frame of aptamer data, typically subsetting
SOMAmer features (columns) to SomaSciences 1129 content.
}
\examples{
dim(sample.adat)
x <- modifyAptContent(sample.adat)
dim(x)
names(SSmenu)
x <- modifyAptContent(sample.adat, menu.ver = "v1.1k_v02")  # v2 1124
dim(x)
}
\seealso{
\code{\link{getSSmenu}}, \code{\link{getSeqIdMatches}}, \code{\link{SSmenu}}
}
\author{
Stu Field
}
