% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.qq.R
\name{my.qq}
\alias{my.qq}
\title{Create Q-Q plots}
\usage{
my.qq(
  y,
  fun = c("norm", "unif", "exp", "binom"),
  na.rm = TRUE,
  col = 2,
  lty = 3,
  lwd = 1.5,
  ...
)
}
\arguments{
\item{y}{The variable of interest as a numeric vector.}

\item{fun}{Which functional distribution you wish to test against. Options
are: \code{"norm"}, \code{"unif"}, \code{"exp"}, and \code{"binom"}.}

\item{na.rm}{Should \code{NA}s be removed from the data prior to analysis?}

\item{col}{Color of the Q-Q line to be plotted in the Q-Q plot.}

\item{lty}{Line type of the Q-Q line to be plotted in the Q-Q plot.}

\item{lwd}{Line width of the Q-Q line to be plotted in the Q-Q plot.}

\item{...}{Additional arguments passed to the \code{fun} function argument
above that describe the shape parameters of the quantile distributions,
typically \code{\link{qnorm}}, \code{\link{qunif}}, \code{\link{qexp}}, and
\code{\link{qbinom}}.}

\item{check.p.values}{Is \code{y} a vector of p.values?}
}
\value{
\item{plot1 }{A histogram of the raw data variable.} \item{plot2 }{A
Q-Q plot of the quantiles of the raw data and the theoretical distribution.}
\item{ShapiroTest }{If \code{fun="norm"}, a Shapiro-Wilk normality test is
performed and the result added to the Q-Q plot.}
}
\description{
General tool for plotting unknown distributions against known distributions,
producing a histogram of the unknown data and a Quantile-Quantile plot of
the relationship.
}
\details{
If \code{fun="norm"}, a Shapiro-Wilk normality test is performed and the
result added to the Q-Q plot.
}
\note{
\%\% ~~further notes~~
}
\examples{

\dontrun{
test = rnorm(500)
my.qq(test, fun="norm")
my.qq(test, fun="unif")
my.qq(test, fun="exp")
test = rexp(500)
my.qq(test, fun="exp")
}

}
\references{
\%\% ~put references to the literature/web site here ~
}
\seealso{
\code{\link{hist}} \cr \code{\link{qnorm}} \cr \code{\link{qunif}}
\cr \code{\link{qexp}} \cr \code{\link{qbinom}}
}
\author{
Stu Field
}
\keyword{~kwd1}
\keyword{~kwd2}
