% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAxes.R
\name{my_x_log10}
\alias{my_x_log10}
\title{Scale x-axis}
\usage{
my_x_log10(expmin = -22, expmax = 22)
}
\arguments{
\item{epxmin}{Numeric. The exponent of the lower limit of labels
and gridlines we are displaying}

\item{epxmax}{Numeric. The exponent of the upper limit of labels
and gridlines we are displaying}
}
\description{
Term to add to ggplots to use with x_log axis (that get nice minor grid lines).
Unless you have a really high or low range, the default of -22 to +22 should be ok.
}
\examples{

df.example <- data.frame(RFU = 10^seq(-5,5,by=0.1))
p <- ggplot(df.example, aes(x=RFU)) + stat_ecdf()
p + my_x_log10()

}
