% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHemo.R
\name{plotHemo}
\alias{plotHemo}
\title{Plot Hemoglobin vs Haptoglobin}
\usage{
plotHemo(
  adat,
  notes.field,
  grep.pattern = NULL,
  cutoff = 30000,
  legend.pos = "bottomleft",
  filename = NULL,
  hb.hp.ratio = 100,
  ...
)
}
\arguments{
\item{adat}{Adat data set of SOMAmer data.}

\item{cutoff}{Where should the cutoff be placed for "high" Hemoglobin measurements.}

\item{legend.pos}{Character. Position of the legend.}

\item{filename}{Optional "filename" if \code{\link[SomaGlobals]{figure}}
is to be implemented.}
}
\value{
A scatter plot and the ADAT rownames of the samples with a ratio
beyond the specified limit.
}
\description{
Makes a scatter plot of all samples for Hemoglobin vs. Haptoglobin.
}
\examples{
plotHemo(sample.adat, hb.hp.ratio = 0.2)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[graphics]{legend}}
}
\author{
Stu Field
}
