% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTwoWayInteraction.R
\name{plotTwoWayInteraction}
\alias{plotTwoWayInteraction}
\title{Plot & Perform two-way ANOVA}
\usage{
plotTwoWayInteraction(
  dat,
  y,
  var,
  levels = list(NULL, NULL),
  shift = 0.25,
  do.int = FALSE,
  y.lab = NULL,
  ...
)
}
\arguments{
\item{dat}{The data frame from which the factor and Response variables are
contained.}

\item{y}{The response variable as a character string.}

\item{var}{A string of \code{length = 2} that contains the first and second
factor variables as they appear in the \code{colnames} of \code{dat}.}

\item{levels}{An optional argument for the ordering of the factors given as
a list; the first entry of the list for the first factor, etc.}

\item{shift}{A numeric value, usually between 0 and 0.5, altering the shift
in the boxes, which will be necessary depending on the factor levels
present.}

\item{do.int}{Logical. Should the p-value for the interaction term be plotted on the
interaction plot?}

\item{...}{Additional arguments passed to \code{boxplot}.}
}
\value{
A summary of the two-way ANOVA table.
}
\description{
Plots an interaction plot (boxplot) of a two-way ANOVA analysis of two
factor variables on a response variable.
}
\examples{
plotTwoWayInteraction(test_data, y = "XYZZ.6969.4.7", var = c("Response", "TimePoint"))
plotTwoWayInteraction(test_data, y = "ABCD.1234.56.8", var = c("TimePoint", "Response"))
plotTwoWayInteraction(test_data, y = "z", var = c("Response", "TimePoint"),
                      levels = list(c("Disease", "Control"), NULL))
}
\seealso{
\code{\link{aov}}, \code{\link{boxplot}}
}
\author{
Stu Field
}
