% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-sink.R
\name{print_sink}
\alias{print_sink}
\title{Print Object to File}
\usage{
print_sink(..., file, width = 250)
}
\arguments{
\item{...}{Object(s) to be printed to file as standard output.}

\item{file}{\code{character(1)}. File name for the output.}

\item{width}{\code{numeric(1)}. Controls the maximum number of
columns on a line used in printing vectors, matrices
and arrays, and when filling by \code{\link[=cat]{cat()}}.
See also \code{getOption("width")}.}
}
\description{
A thin wrapper around the \code{\link[=sink]{sink()}} function to print
the corresponding object to a file connection.
}
\examples{
tab <- cross_tabulate(mtcars, cyl, carb)
print_sink(tab, file = "table_file.txt")
}
