% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-columns.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove Column Containing Specified Entry(s)}
\usage{
remove_columns(x, search)
}
\arguments{
\item{x}{A matrix or data frame. Any 2-dimensional object in R.}

\item{search}{The exact match to be searched.
If contained within any column, it will be removed.}
}
\value{
Same class object as \code{x} with possibly column(s) removed.
}
\description{
Remove an entire column of a data table whose
entries contain a specified value. To remove
\emph{rows} with a matching pattern, use \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\examples{
M <- matrix(1:25, ncol = 5L)
M
remove_columns(M, search = 20)
N <- data.frame(M)
N[4L, 4L] <- "A"
N
remove_columns(N, search = c("A", 17, 23))
}
\author{
Stu Field
}
