% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustPCA.R
\name{robustPCA}
\alias{robustPCA}
\title{Perform Robust PCA}
\usage{
robustPCA(x, tolerance = 1e-07, max.iter = 250, verbose = getOption("verbose"))
}
\arguments{
\item{x}{A data matrix, typically of RFU data. No meta data.}

\item{tolerance}{The tolerance for the error of the Euclidean norm
of the matrix used to stop the iterations. Iterations are maxed at 250
by default.}

\item{max.iter}{Integer. Maximum number of iterations to perform.}

\item{verbose}{Logical. Should iteration "Rounds" be displayed at
the console as the fitting progresses? Also determines whether a plot
of the matrix error in the Euclidean norm decreases as the
iterations progress.}
}
\value{
A list of:
\item{L }{A matrix same dimensions as \code{x} and \emph{very} close values.}
\item{S }{Shrinkage matrix? At any rate, a \emph{very} sparse matrix.}
}
\description{
Need more description here, perhaps consult Mike Mehan on
the reference and how this procedure is justified and validated.
}
\examples{
data <- data.matrix(sample.adat[, getAptamers(sample.adat)])
pca1 <- robustPCA(data)    # basic

# Run until tolerance is reached & plot
pca2 <- robustPCA(data, max.iter = 1000, verbose = TRUE)
}
\author{
Mike Mehan
}
