\name{saveRLObject-methods}
\docType{methods}
\alias{saveRLObject-methods}
\alias{saveRLObject,RLBigData-method}
\alias{saveRLObject,RLResult-method}
\title{Methods for Function saveRLObject}
\description{
Methods for function \code{saveRLObject}. See \link{saveRLObject}
for details.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "RLBigData")}}{

}

\item{\code{signature(object = "RLResult")}}{

}
}}
\keyword{methods}
\keyword{classif}
