% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambleKS.R
\name{scrambleKS}
\alias{scrambleKS}
\alias{plot.scramble.ks}
\title{Scramble KS-distances}
\usage{
scrambleKS(
  data,
  apts,
  response = "Response",
  nperm = 50,
  do.plot = FALSE,
  random.seed = sample(1000, 1),
  verbose = getOption("verbose"),
  MC = FALSE,
  cores = 7,
  ...
)

\method{plot}{scramble.ks}(x, plot.range = NULL, plot.title = "KS distances", ...)
}
\arguments{
\item{data}{A data frame containing SOMAmer data and a "Response" column, preferably
a "training.data" object, see \code{\link{createTrainingData}}.}

\item{apts}{A vector character string of aptamers to be considered in the analysis.}

\item{response}{Character. String of the column name to be scrambled ... defaults
to the \code{Response} column of the data frame (ideal for "training.data" objects).}

\item{nperm}{Integer. The number of bootstrap trials to perform when calculating
the confidence intervals (95\\%).}

\item{do.plot}{Should a plot be created of the scrambled data?}

\item{random.seed}{Numeric. Optional random seed to set the permutations for
reproducible scrambling.}

\item{MC}{Logical. Should multicore processing be used? Linux/Mac only.}

\item{cores}{Integer. If \code{MC=TRUE}, the number of cores to use.}

\item{...}{Additional arguments passed to \code{\link{plot.scramble.ks}}
OR additional argumnets passed to \code{\link[graphics]{barplot}} (if called
during the \code{plot} method.}

\item{x}{Data object of class "scramble.ks" when using the S3 plot method.}

\item{plot.range}{A range (or subset) of features to plot (rather than
plotting the entire array.}

\item{plot.title}{Character. A string for the plot title.}

\item{title}{Title label for the plot.}
}
\value{
An object of class \code{"scramble.ks"}.
}
\description{
Class scrambling of the \code{Response} variable and recalculate
KS-distances for an empirical bootstrap of the likelihood of acheiving same
KS scores. Bars represent 95\\% confidence interval about the unscrambled
KS-distance for each aptamer.
}
\section{Functions}{
\itemize{
\item \code{plot(scramble.ks)}: Plot Scrambled KS-distances. An S3 plot method for objects of class "scramble.ks".

}}
\examples{
trdat <- convert2TrainingData(sample.adat, "SampleGroup")
perm <- scrambleKS(trdat, apts=getAptamers(trdat), verb=TRUE, random.seed=101) # 30s
perm <- scrambleKS(trdat, apts=getAptamers(trdat), random.seed=101, MC=TRUE)   # 06s
# S3 plot method
plot(perm, plot.range = 1:15)
}
\seealso{
\code{\link[graphics]{barplot}}, \code{\link{scramble.ks}}
}
\author{
Stu Field
}
