% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriateHeatmap.R
\name{seriateHeatmap}
\alias{seriateHeatmap}
\title{Heatmap Seriation}
\usage{
seriateHeatmap(data, method = "OLO", censor.crit = FALSE)
}
\arguments{
\item{data}{The "soma.adat" object containing aptamer data.}

\item{method}{Which clustering method to use, see \code{\link[seriation]{seriate}}.}

\item{censor.crit}{Logical. Should the correlation values be cut off at a
critical value.}
}
\value{
A seriated heatmap of the entire data frame.
}
\description{
Produce a heatmap of all SOMAmers in an adat. Just a black and white total
heatmap of ordered, correlated analytes.
}
\examples{
seriateHeatmap(sample.adat)
}
\author{
Stu Field
}
