% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_test.R
\name{sign_test}
\alias{sign_test}
\title{Calculate Sign Test}
\usage{
sign_test(x, y, ...)
}
\arguments{
\item{x}{First vector of values, in order of individual.}

\item{y}{Second vector of values, in order of individual,
to be paired with \code{x}.}

\item{...}{Additional arguments passed to \code{\link[stats]{binom.test}}.}
}
\value{
The result of \code{\link[stats]{binom.test}} on the signs.
}
\description{
Calculates a simple sign test for paired data
whether the number of "ups" is equal to the number
of "downs". Tests the null probability \code{p = 0.5}.
}
\examples{
sign.test(rnorm(10), rnorm(10))
vec1 <- rnorm(20)
s <- sample(1:20, 15)
vec2 <- vec1
vec2[s] <- vec2[s] + 1
vec2[setdiff(1:20, s)] <- vec2[setdiff(1:20, s)] - 1
sign.test(vec1, vec2)
}
\references{
Mick Crawley, The R Book, pg. 300.
}
\seealso{
\code{\link[stats]{binom.test}}
}
\author{
Stu Field
}
