% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sjt.mlm}
\alias{sjt.mlm}
\title{Show linear regression with mediator variable as HTML table}
\usage{
sjt.mlm(iv, dv, mv, controls = NULL, data)
}
\description{
Shows fitted linear models with mediator effects
               as HTML table, or saves them as file.
}
\seealso{
\code{\link{sjt.glm}} \cr
         \code{\link{sjp.lm}} \cr
         \code{\link{sjp.lm1}} \cr
         \code{\link{sjt.lm}}
}

