% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sju.groupString}
\alias{sju.groupString}
\title{Group near elements of string vectors}
\usage{
sju.groupString(strings, maxdist)
}
\arguments{
\item{strings}{a character vector with string elements}

\item{maxdist}{the maximum distance between two string elements, which is allowed to treat two
         elements as similar or equal.}
}
\value{
A character vector where similar string elements (values) are recoded into a new, single value.
}
\description{
This function groups elements of a string vector (character or string variable) according
               to the element's distance. The more similar two string elements are, the higher is the
               chance to be combined into a group.
}
\examples{
oldstring <- c("Hello", "Helo", "Hole", "Apple", "Ape", "New", "Old", "System", "Systemic")
table(oldstring)
newstring <- sju.groupString(oldstring, 3)
newstring
table(newstring)
}

