% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl.R
\name{sl}
\alias{sl}
\title{Source Library of Functions}
\usage{
sl(file = "func_lib.R", name = "lib")
}
\description{
Source a library of functions from a file
into a local environment and attach it.
}
\examples{
sl("analysis-functions.R")
}
\seealso{
\code{\link[base]{sys.source}}
}
\author{
Stu Field
}
