% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys.sourceDir.R
\name{Sys.sourceDir}
\alias{Sys.sourceDir}
\title{Source Files from Directory}
\usage{
Sys.sourceDir(path, envir.name = "e")
}
\arguments{
\item{path}{Location of the directory containing \verb{*.R} files.}

\item{envir.name}{Character. A name for the environment to
place the function definitions. This name will be used
when attaching the environment to the search path.}
}
\description{
Evaluates and parses all \verb{*.R} files and their contained functions
into a specified environment and attaches that environment
to the search path.
}
\note{
The environment is attached in the \emph{final} position
in the search path.
}
\examples{
\dontrun{
dir <- paste0(Sys.getenv("R_SOMA_DEV"), "sandbox", "/R")
Sys.sourceDir(dir, envir.name = "sandbox")
}
}
\seealso{
\code{\link{sys.source}}, \code{\link[fs]{dir_ls}}, \code{\link{attach}}
}
\author{
Stu Field
}
