% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-case.R
\name{title_case}
\alias{title_case}
\title{Capitalize Title Format}
\usage{
title_case(x, strict = FALSE)
}
\arguments{
\item{x}{A character string in the form of a sentence
to be converted to "title case" (i.e. first letter capitalized).}

\item{strict}{Logical. Should first letter capitalization
be \emph{strictly} applied? See example.}
}
\value{
A character string with "title case" conversion.
}
\description{
Change case to capitalize first letter of
each word in a character string.
}
\examples{
title_case("using AIC for model selection")
## ->  [1] "Using AIC For Model Selection"
title_case(c("using AIC", "for MODEL selection"), strict = TRUE)
## ->    [1] "Using Aic"  "For Model Selection"
##                  ^^^        ^^^^^
##                 "bad"       "good"
}
\seealso{
\code{\link[=toupper]{toupper()}}, \code{\link[=tolower]{tolower()}}, \code{\link[=strsplit]{strsplit()}}
}
\author{
Stu Field
}
