% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples_test.R
\name{triples_test}
\alias{triples_test}
\title{Triples Test}
\usage{
triples_test(x, alpha = 0.05)
}
\arguments{
\item{x}{\code{numeric(n)}. A vector of values.}

\item{alpha}{\code{numeric(1)}. The significance level.}
}
\value{
return_value
}
\description{
Doesn't quite work right yet ...
}
\note{
Ho: null hypothesis is that the data are symmetric:
\code{x = c(17.4, 17.9, 17.6, 18.1, 17.6)}
}
\examples{
triples_test(rnorm(10))
}
\references{
From Hollander & Wolfe (1999); NonParametric Statistical Methods, pg 87-91.
}
\author{
Stu Field
}
